fixing warnings about using non-integer for integer

+++ ad_clicks.py
@@ -42,7 +42,7 @@ def writeAdClicksCSV(startTime, dayDuration):
 	#pick 0-30% users for clicking based on their click probabilities
 	factor = random.uniform(0, 0.3)
 	#print factor
-	adUsers = np.random.choice(range(0, len(totalUsers)), factor*len(totalUsers), replace=True, p=adProbabilities).tolist()
+	adUsers = np.random.choice(range(0, len(totalUsers)), int(factor*len(totalUsers)), replace=True, p=adProbabilities).tolist()
 	adclicks = []
 
 	#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~GENERATE adclicks from these users
